hotfix: remove derive functions from common/crypto
